Pull broadcastRecoveryAddress into its own file
